requirements.txt
Django==3.1.1
django-filter
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
PyJWT==1.7.1
requests==2.23.0